# watchlist/backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Smart Market Watchlist",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/watchlist/digest": {
      "get": {
        "tags": [
          "watchlist"
        ],
        "summary": "Digest",
        "operationId": "digest_api_watchlist_digest_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DigestOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/watchlist/briefing": {
      "get": {
        "tags": [
          "watchlist"
        ],
        "summary": "Briefing",
        "operationId": "briefing_api_watchlist_briefing_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BriefingOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/watchlist/items": {
      "post": {
        "tags": [
          "watchlist"
        ],
        "summary": "Add Item",
        "operationId": "add_item_api_watchlist_items_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ItemAdd"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Item"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/watchlist/items/{symbol}": {
      "delete": {
        "tags": [
          "watchlist"
        ],
        "summary": "Remove Item",
        "operationId": "remove_item_api_watchlist_items__symbol__delete",
        "parameters": [
          {
            "name": "symbol",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Symbol"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/watchlist/seen": {
      "post": {
        "tags": [
          "watchlist"
        ],
        "summary": "Mark Seen",
        "operationId": "mark_seen_api_watchlist_seen_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SeenIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SeenOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/rules/compile": {
      "post": {
        "tags": [
          "rules"
        ],
        "summary": "Compile Text",
        "operationId": "compile_text_api_rules_compile_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RuleCompileIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuleCompileOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/rules": {
      "post": {
        "tags": [
          "rules"
        ],
        "summary": "Create Rule",
        "operationId": "create_rule_api_rules_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RuleCreateIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuleOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "rules"
        ],
        "summary": "List Rules",
        "operationId": "list_rules_api_rules_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RuleListItem"
                  },
                  "title": "Response List Rules Api Rules Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/rules/{rule_id}": {
      "delete": {
        "tags": [
          "rules"
        ],
        "summary": "Delete Rule",
        "operationId": "delete_rule_api_rules__rule_id__delete",
        "parameters": [
          {
            "name": "rule_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Rule Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BriefingOut": {
        "properties": {
          "text": {
            "type": "string",
            "maxLength": 600,
            "title": "Text"
          },
          "source": {
            "type": "string",
            "enum": [
              "llm",
              "template"
            ],
            "title": "Source"
          },
          "generated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Generated At"
          },
          "was_cached": {
            "type": "boolean",
            "title": "Was Cached"
          }
        },
        "type": "object",
        "required": [
          "text",
          "source",
          "generated_at",
          "was_cached"
        ],
        "title": "BriefingOut"
      },
...
      "DigestOut": {
        "properties": {
          "now": {
            "type": "string",
            "format": "date-time",
            "title": "Now"
          },
          "market_status": {
            "type": "string",
            "enum": [
              "open",
              "closed",
              "pre_open"
            ],
            "title": "Market Status"
          },
          "replay_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Replay Date"
          },
          "latest_bar_date": {
            "type": "string",
            "format": "date",
            "title": "Latest Bar Date"
          },
          "away_duration_seconds": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Away Duration Seconds"
          },
          "last_reviewed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Reviewed At"
          },
          "changed_count": {
            "type": "integer",
            "title": "Changed Count"
          },
          "total_count": {
            "type": "integer",
            "title": "Total Count"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/Item"
            },
            "type": "array",
            "title": "Items"
          },
          "providers_degraded": {
            "type": "boolean",
            "title": "Providers Degraded"
          }
        },
        "type": "object",
        "required": [
          "now",
          "market_status",
          "replay_date",
          "latest_bar_date",
          "away_duration_seconds",
          "last_reviewed_at",
          "changed_count",
          "total_count",
          "items",
          "providers_degraded"
        ],
        "title": "DigestOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Item": {
        "properties": {
          "symbol": {
            "type": "string",
            "pattern": "^[A-Z0-9&-]{1,20}\\.(NS|BO)$",
            "title": "Symbol"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "industry": {
            "type": "string",
            "title": "Industry"
          },
          "quote": {
            "$ref": "#/components/schemas/Quote"
          },
          "today_change_pct": {
            "type": "number",
            "title": "Today Change Pct"
          },
          "peer_change_pct": {
            "type": "number",
            "title": "Peer Change Pct"
          },
          "residual_pct": {
            "type": "number",
            "title": "Residual Pct"
          },
          "z_score": {
            "type": "number",
            "title": "Z Score"
          },
          "raw_z_score": {
            "type": "number",
            "title": "Raw Z Score"
          },
          "rvol": {
            "type": "number",
            "title": "Rvol"
          },
          "rvol_is_approximate": {
            "type": "boolean",
            "title": "Rvol Is Approximate"
          },
          "change_since_seen_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Change Since Seen Pct"
          },
          "last_seen_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Seen At"
          },
          "attention": {
            "type": "string",
            "enum": [
              "high",
              "notable",
              "quiet"
            ],
            "title": "Attention"
          },
          "is_changed": {
            "type": "boolean",
            "title": "Is Changed"
          },
          "low_confidence": {
            "type": "boolean",
            "title": "Low Confidence"
          },
          "signals": {
            "items": {
              "$ref": "#/components/schemas/Signal"
            },
            "type": "array",
            "title": "Signals"
          },
          "levels": {
            "$ref": "#/components/schemas/Levels"
          },
          "sma_distance_pct": {
            "$ref": "#/components/schemas/SmaDistance"
          },
          "peer": {
            "$ref": "#/components/schemas/Peer"
          },
          "catalyst_status": {
            "type": "string",
            "enum": [
              "not_fetched",
              "pending",
              "found",
              "none_found",
              "unavailable"
            ],
            "title": "Catalyst Status"
          }
        },
        "type": "object",
        "required": [
          "symbol",
          "name",
          "industry",
          "quote",
          "today_change_pct",
          "peer_change_pct",
          "residual_pct",
          "z_score",
          "raw_z_score",
          "rvol",
          "rvol_is_approximate",
          "change_since_seen_pct",
          "last_seen_at",
          "attention",
          "is_changed",
          "low_confidence",
          "signals",
          "levels",
          "sma_distance_pct",
          "peer",
          "catalyst_status"
        ],
        "title": "Item"
      },
      "ItemAdd": {
        "properties": {
          "symbol": {
            "type": "string",
            "pattern": "^[A-Z0-9&-]{1,20}\\.(NS|BO)$",
            "title": "Symbol"
          }
        },
        "type": "object",
        "required": [
          "symbol"
        ],
        "title": "ItemAdd"
      },
      "Levels": {
        "properties": {
          "high_52w": {
            "type": "number",
            "title": "High 52W"
          },
          "low_52w": {
            "type": "number",
            "title": "Low 52W"
          },
          "prev_high": {
            "type": "number",
            "title": "Prev High"
          },
          "prev_low": {
            "type": "number",
            "title": "Prev Low"
          }
        },
        "type": "object",
        "required": [
          "high_52w",
          "low_52w",
          "prev_high",
          "prev_low"
        ],
        "title": "Levels"
      },
...
      "Quote": {
        "properties": {
          "price": {
            "type": "number",
            "title": "Price"
          },
          "prev_close": {
            "type": "number",
            "title": "Prev Close"
          },
          "day_high": {
            "type": "number",
            "title": "Day High"
          },
          "day_low": {
            "type": "number",
            "title": "Day Low"
          },
          "volume": {
            "type": "integer",
            "title": "Volume"
          },
          "as_of": {
            "type": "string",
            "format": "date-time",
            "title": "As Of"
          },
          "source": {
            "type": "string",
            "enum": [
              "yahoo",
              "bse"
            ],
            "title": "Source"
          },
          "staleness_seconds": {
            "type": "integer",
            "title": "Staleness Seconds"
          },
          "confidence": {
            "type": "string",
            "enum": [
              "fresh",
              "delayed",
              "stale",
              "disputed",
              "closed"
            ],
            "title": "Confidence"
          },
          "alt": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AltQuote"
              },
              {
                "type": "null"
              }
            ]
          },
          "divergence_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Divergence Pct"
          }
        },
        "type": "object",
        "required": [
          "price",
          "prev_close",
          "day_high",
          "day_low",
          "volume",
          "as_of",
          "source",
          "staleness_seconds",
          "confidence",
          "alt",
          "divergence_pct"
        ],
        "title": "Quote"
      },
...
      "RuleCompileIn": {
        "properties": {
          "text": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Text"
          }
        },
        "type": "object",
        "required": [
          "text"
        ],
        "title": "RuleCompileIn"
      },
      "RuleCompileOut": {
        "properties": {
          "rule": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Rule"
              },
              {
                "type": "null"
              }
            ]
          },
          "preview": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preview"
          },
          "error": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "rule",
          "preview"
        ],
        "title": "RuleCompileOut"
      },
...
      "RuleCreateIn": {
        "properties": {
          "nl_text": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Nl Text"
          },
          "rule": {
            "$ref": "#/components/schemas/Rule"
          }
        },
        "type": "object",
        "required": [
          "nl_text",
          "rule"
        ],
        "title": "RuleCreateIn"
      },
      "RuleListItem": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "nl_text": {
            "type": "string",
            "title": "Nl Text"
          },
          "preview": {
            "type": "string",
            "title": "Preview"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "matched_today": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Matched Today"
          }
        },
        "type": "object",
        "required": [
          "id",
          "nl_text",
          "preview",
          "enabled",
          "created_at",
          "matched_today"
        ],
        "title": "RuleListItem"
      },
      "RuleOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "nl_text": {
            "type": "string",
            "title": "Nl Text"
          },
          "rule": {
            "$ref": "#/components/schemas/Rule"
          },
          "preview": {
            "type": "string",
            "title": "Preview"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "nl_text",
          "rule",
          "preview",
          "enabled",
          "created_at"
        ],
        "title": "RuleOut"
      },
...
      "SeenIn": {
        "properties": {
          "symbols": {
            "anyOf": [
              {
                "items": {
                  "type": "string",
                  "pattern": "^[A-Z0-9&-]{1,20}\\.(NS|BO)$"
                },
                "type": "array"
              },
              {
                "type": "string",
                "const": "all"
              }
            ],
            "maxLength": 100,
            "title": "Symbols"
          }
        },
        "type": "object",
        "required": [
          "symbols"
        ],
        "title": "SeenIn"
      },
      "SeenOut": {
        "properties": {
          "marked": {
            "type": "integer",
            "title": "Marked"
          },
          "reviewed_at": {
            "type": "string",
            "format": "date-time",
            "title": "Reviewed At"
          }
        },
        "type": "object",
        "required": [
          "marked",
          "reviewed_at"
        ],
        "title": "SeenOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}